Move GC to its own file
